+++ src/us/kbase/workspace/database/ByteArrayFileCacheManager.java
@@ -83,6 +83,7 @@ public class ByteArrayFileCacheManager {
 			if (tfm == null) {
 				final ByteArrayOutputStream baos = new ByteArrayOutputStream();
 				final int size = IOUtils.copy(input, baos);
+				// TODO CODE this creates a new byte[], doubling the memory usage temporarily
 				@SuppressWarnings("resource")
 				final JsonTokenStream jts = new JsonTokenStream(baos.toByteArray());
 				return new ByteArrayFileCache(
@@ -131,6 +132,7 @@ public class ByteArrayFileCacheManager {
 			if (tfm == null) {
 				final ByteArrayOutputStream baos = new ByteArrayOutputStream();
 				parent.getSubdataExtractionAsStream(paths, baos);
+				// TODO CODE this creates a new byte[], doubling the memory usage temporarily
 				@SuppressWarnings("resource")
 				final JsonTokenStream jts = new JsonTokenStream(baos.toByteArray());
 				return new ByteArrayFileCache(