+++ src/us/kbase/workspace/test/kbase/SchemaUpdaterCLITest.java
@@ -41,7 +41,7 @@ import us.kbase.workspace.kbase.InitWorkspaceServer.WorkspaceInitException;
 
 public class SchemaUpdaterCLITest {
 	
-	private static final String VERSION = "0.13.0-dev2";
+	private static final String VERSION = "0.13.0-dev3";
 	
 	private static final List<String> HELP = list(
 			"Usage: update_workspace_database_schema [-chosV] <CONFIG_FILE>",