+++ docsource/releasenotes.rst
@@ -24,6 +24,7 @@ UPDATES:
 * The sample service ACL update mechanism has been changed to use a bulk interface for updating
   ACLs rather than updating ACLs one at a time, which should speed up fetching objects with
   ``@id sample`` annotations.
+* The ``batch_external_system_updates`` option has been added to ``get_objects2``.
 
 VERSION: 0.12.1 (Released 1/25/2022)
 ------------------------------------
@@ -31,7 +32,7 @@ VERSION: 0.12.1 (Released 1/25/2022)
 UPDATES: 
 
 * The ``skip_external_system_updates`` option has been added to ``get_objects2``.
-* Added the ``--print-stacktrace`` option to the schema updater
+* Added the ``--print-stacktrace`` option to the schema updater.
 
 VERSION: 0.12.0 (Released 11/16/21)
 -----------------------------------